src/main/java/gift/product/dao/MemoryProductRepository.java
package gift.product.dao;

import gift.product.domain.Product;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import java.util.Optional;

public class MemoryProductRepository implements ProductRepository {

    private final Map<Long, Product> products = new HashMap<>();

    @Override
    public Product save(Product product) {
        products.put(product.getId(), product);
        return products.get(product.getId());
    }

    @Override
    public Optional<Product> findById(Long id) {
        return Optional.ofNullable(products.get(id));
    }

    @Override
    public List<Product> findAll() {
        return new ArrayList<>(products.values());
    }

    @Override
    public void deleteById(Long id) {
        products.remove(id);
    }

    @Override
    public void deleteAll() {
        products.clear();
    }

    @Override
    public int update(Long id, Product product) {
        return Math.toIntExact(products.put(id, product).getId());
    }

}
